Refactor JsonDatabaseMigration to make it easier to use
* This refactor make possible to implement a JsonDatabaseMigration only defining the setting name
and the list of JSON fields to update/modify.
